+++ CHANGELOG.md
@@ -9,7 +9,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Fixed persistent modal
+- Added a frame for the event window
 - fixed node label name in YGraph
+- fixed the initial position of dash
+- fixed dash z index
+- fixed tooltip z-index
+
 
 ## [1.81.0]